refactor(compaction): keep review surfaces focused
