+++ internal/message/compaction_epoch_postgres_integration_test.go
@@ -103,15 +103,15 @@ func TestPostgresStaleCompactionAttemptCannotCompleteOK(t *testing.T) {
 	}
 	assertSessionCompactionEpoch(t, ctx, tx, 1)
 	if _, err := queries.CreateCompactionLog(ctx, dbsqlc.CreateCompactionLogParams{
-		BotID:        mustTestUUID(t, postgresMessageTestBotID),
-		SessionID:    mustTestUUID(t, postgresMessageTestSessionID),
+		BotID:         mustTestUUID(t, postgresMessageTestBotID),
+		SessionID:     mustTestUUID(t, postgresMessageTestSessionID),
 		ExpectedEpoch: 0,
 	}); !errors.Is(err, pgx.ErrNoRows) {
 		t.Fatalf("create compaction log with stale epoch = %v, want pgx.ErrNoRows", err)
 	}
 	if _, err := queries.CreateCompactionLog(ctx, dbsqlc.CreateCompactionLogParams{
-		BotID:        mustTestUUID(t, postgresMessageTestBotID),
-		SessionID:    mustTestUUID(t, postgresMessageTestSessionID),
+		BotID:         mustTestUUID(t, postgresMessageTestBotID),
+		SessionID:     mustTestUUID(t, postgresMessageTestSessionID),
 		ExpectedEpoch: 1,
 	}); err != nil {
 		t.Fatalf("create compaction log with current epoch: %v", err)
